+++ src/lib/builtins/array.rs
@@ -128,9 +128,7 @@ pub fn pop(this: &Value, _: &[Value], _: &mut Interpreter) -> ResultValue {
     let curr_length: i32 =
         from_value(this.get_field_slice("length")).expect("Could not convert argument to i32");
     if curr_length < 1 {
-        return Err(to_value(
-            "Cannot pop() on an array with zero length".to_string(),
-        ));
+        return Ok(Gc::new(ValueData::Undefined));
     }
     let pop_index = curr_length.wrapping_sub(1);
     let pop_value: Value = this.get_field(&pop_index.to_string());
@@ -507,6 +505,26 @@ pub fn fill(this: &Value, args: &[Value], _: &mut Interpreter) -> ResultValue {
     Ok(this.clone())
 }
 
+pub fn includes_value(this: &Value, args: &[Value], _: &mut Interpreter) -> ResultValue {
+    let search_element = args
+        .get(0)
+        .cloned()
+        .unwrap_or_else(|| Gc::new(ValueData::Undefined));
+
+    let length: i32 =
+        from_value(this.get_field_slice("length")).expect("Could not get `length` property.");
+
+    for idx in 0..length {
+        let check_element = this.get_field(&idx.to_string()).clone();
+
+        if check_element == search_element {
+            return Ok(to_value(true));
+        }
+    }
+
+    Ok(to_value(false))
+}
+
 /// Create a new `Array` object
 pub fn create_constructor(global: &Value) -> Value {
     // Create Constructor
@@ -531,6 +549,8 @@ pub fn create_constructor(global: &Value) -> Value {
     index_of_func.set_field_slice("length", to_value(1_i32));
     let last_index_of_func = to_value(last_index_of as NativeFunctionData);
     last_index_of_func.set_field_slice("length", to_value(1_i32));
+    let includes_func = to_value(includes_value as NativeFunctionData);
+    includes_func.set_field_slice("length", to_value(1_i32));
 
     array_prototype.set_field_slice("push", push_func);
     array_prototype.set_field_slice("pop", to_value(pop as NativeFunctionData));
@@ -541,6 +561,7 @@ pub fn create_constructor(global: &Value) -> Value {
     array_prototype.set_field_slice("every", to_value(every as NativeFunctionData));
     array_prototype.set_field_slice("find", to_value(find as NativeFunctionData));
     array_prototype.set_field_slice("findIndex", to_value(find_index as NativeFunctionData));
+    array_prototype.set_field_slice("includes", includes_func);
     array_prototype.set_field_slice("indexOf", index_of_func);
     array_prototype.set_field_slice("lastIndexOf", last_index_of_func);
     array_prototype.set_field_slice("fill", to_value(fill as NativeFunctionData));
@@ -687,6 +708,95 @@ mod tests {
 
         let missing = forward(&mut engine, "missing.findIndex(comp)");
         assert_eq!(missing, String::from("-1"));
+    }
+
+    #[test]
+    fn push() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+        let init = r#"
+        var arr = [1, 2];
+        "#;
+        forward(&mut engine, init);
+
+        assert_eq!(forward(&mut engine, "arr.push()"), "2");
+        assert_eq!(forward(&mut engine, "arr.push(3, 4)"), "4");
+        assert_eq!(forward(&mut engine, "arr[2]"), "3");
+        assert_eq!(forward(&mut engine, "arr[3]"), "4");
+    }
+
+    #[test]
+    fn pop() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+        let init = r#"
+        var empty = [ ];
+        var one = [1];
+        var many = [1, 2, 3, 4];
+        "#;
+        forward(&mut engine, init);
+
+        assert_eq!(
+            forward(&mut engine, "empty.pop()"),
+            String::from("undefined")
+        );
+        assert_eq!(forward(&mut engine, "one.pop()"), "1");
+        assert_eq!(forward(&mut engine, "one.length"), "0");
+        assert_eq!(forward(&mut engine, "many.pop()"), "4");
+        assert_eq!(forward(&mut engine, "many[0]"), "1");
+        assert_eq!(forward(&mut engine, "many.length"), "3");
+    }
+
+    #[test]
+    fn shift() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+        let init = r#"
+        var empty = [ ];
+        var one = [1];
+        var many = [1, 2, 3, 4];
+        "#;
+        forward(&mut engine, init);
+
+        assert_eq!(
+            forward(&mut engine, "empty.shift()"),
+            String::from("undefined")
+        );
+        assert_eq!(forward(&mut engine, "one.shift()"), "1");
+        assert_eq!(forward(&mut engine, "one.length"), "0");
+        assert_eq!(forward(&mut engine, "many.shift()"), "1");
+        assert_eq!(forward(&mut engine, "many[0]"), "2");
+        assert_eq!(forward(&mut engine, "many.length"), "3");
+    }
+
+    #[test]
+    fn unshift() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+        let init = r#"
+        var arr = [3, 4];
+        "#;
+        forward(&mut engine, init);
+
+        assert_eq!(forward(&mut engine, "arr.unshift()"), "2");
+        assert_eq!(forward(&mut engine, "arr.unshift(1, 2)"), "4");
+        assert_eq!(forward(&mut engine, "arr[0]"), "1");
+        assert_eq!(forward(&mut engine, "arr[1]"), "2");
+    }
+
+    #[test]
+    fn reverse() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+        let init = r#"
+        var arr = [1, 2];
+        var reversed = arr.reverse();
+        "#;
+        forward(&mut engine, init);
+        assert_eq!(forward(&mut engine, "reversed[0]"), "2");
+        assert_eq!(forward(&mut engine, "reversed[1]"), "1");
+        assert_eq!(forward(&mut engine, "arr[0]"), "2");
+        assert_eq!(forward(&mut engine, "arr[1]"), "1");
     }
 
     #[test]
@@ -893,5 +1003,44 @@ mod tests {
         forward(&mut engine, "a = (new Array(3)).fill({});");
         forward(&mut engine, "a[0].hi = 'hi';");
         assert_eq!(forward(&mut engine, "a[1].hi"), String::from("hi"));
+    }
+
+    #[test]
+    fn inclues_value() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+        let init = r#"
+        var empty = [ ];
+        var one = ["a"];
+        var many = ["a", "b", "c"];
+        var duplicates = ["a", "b", "c", "a", "b"];
+        var undefined = [undefined];
+        "#;
+        forward(&mut engine, init);
+
+        // Empty
+        let empty = forward(&mut engine, "empty.includes('a')");
+        assert_eq!(empty, String::from("false"));
+
+        // One
+        let one = forward(&mut engine, "one.includes('a')");
+        assert_eq!(one, String::from("true"));
+        // Missing from one
+        let missing_from_one = forward(&mut engine, "one.includes('b')");
+        assert_eq!(missing_from_one, String::from("false"));
+
+        // In many
+        let first_in_many = forward(&mut engine, "many.includes('c')");
+        assert_eq!(first_in_many, String::from("true"));
+        // Missing from many
+        let second_in_many = forward(&mut engine, "many.includes('d')");
+        assert_eq!(second_in_many, String::from("false"));
+
+        // In duplicates
+        let first_in_many = forward(&mut engine, "duplicates.includes('a')");
+        assert_eq!(first_in_many, String::from("true"));
+        // Missing from duplicates
+        let second_in_many = forward(&mut engine, "duplicates.includes('d')");
+        assert_eq!(second_in_many, String::from("false"));
     }
 }